Alerts,alerts Dialogs and error boundary added

+++ frontend/src/screens/RemoveBgScreen.jsx
@@ -3,7 +3,9 @@ import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 
-import { fetchRemoveBg, fetchGetRemoveBg } from '@/features/RemoveBgSlice'
+import { fetchRemoveBg, fetchGetRemoveBg, resetRemoveBg } from '@/features/RemoveBgSlice'
+import ServerError from '@/components/ServerError'
+import CustomAlert from '@/components/CustomAlert'
 
 import ReactCompareImage from 'react-compare-image'
 import { Loader2 } from "lucide-react"
@@ -34,14 +36,16 @@ function RemoveBgScreen() {
 
     const [hide, setHide] = useState(false)
     const [model, setModel] = useState('')
+    const [isDragOver, setIsDragOver] = useState(false);
 
-    const is_varified = userInfo ? userInfo.is_varified : false
+    const is_verified = userInfo ? userInfo.is_verified : false
 
     useEffect(() => {
-        if (!userInfo || !is_varified) {
+        if (!is_verified) {
             navigate('/login')
         }
     }, [userInfo, navigate])
+
 
     useEffect(() => {
         if (removeBgStatus === 'succeeded') {
@@ -52,14 +56,17 @@ function RemoveBgScreen() {
     const handleDrop = (e) => {
         e.preventDefault();
         const file = e.dataTransfer.files[0];
-        console.log(file);
         if (file.type.startsWith('image/')) {
             dispatch(fetchRemoveBg({
                 model: model,
                 image: file
             }))
         } else {
-            alert('Please drop an image file.');
+            setIsDragOver(true);
+            const timer = setTimeout(() => {
+                setIsDragOver(false);
+            }, 3700);
+            return () => clearTimeout(timer);
         }
     };
     const handleDragOver = (e) => {
@@ -67,10 +74,20 @@ function RemoveBgScreen() {
     };
 
     const uploadHndler = (e) => {
-        dispatch(fetchRemoveBg({
-            model: model,
-            image: e.target.files[0]
-        }))
+        e.preventDefault();
+        const file = e.target.files[0];
+        if (file.type.startsWith('image/')) {
+            dispatch(fetchRemoveBg({
+                model: model,
+                image: file
+            }))
+        } else {
+            setIsDragOver(true);
+            const timer = setTimeout(() => {
+                setIsDragOver(false);
+            }, 3700);
+            return () => clearTimeout(timer);
+        }
     }
 
     const animeHandler = () => {
@@ -88,64 +105,81 @@ function RemoveBgScreen() {
         setHide(true)
     }
 
-    return (
-        <div className='w-full mx-auto flex justify-center items-center'>
-            <Card className='w-[95%] md:w-[80%] lg:w-[60%] mt-10'>
-                <CardHeader>
-                    <CardTitle className="text-lg md:text-2xl text-center">Remove Background</CardTitle>
-                </CardHeader>
-                <CardContent>
-                    {removeBgStatus === 'idle' ? (
-                        <div className="flex flex-col space-y-4 my-2 items-center ">
-                            {!hide && (
-                                <div className="flex flex-col items-center space-y-2 text-sm md:text-base">
-                                    <p>Before uploading the image choose the image type if both are not working use other</p>
-                                    <div className='grid grid-cols-2 md:grid-cols-3 gap-2' >
-                                        <Button variant="outline" onClick={animeHandler}>Anime</Button>
-                                        <Button variant="outline" onClick={generalHandler}>General</Button>
-                                        <Button variant="outline" onClick={otherHandler}>Other</Button>
+    const resetHandler = () => {
+        dispatch(resetRemoveBg())
+        setHide(false)
+        setModel('')
+    }
 
+    return (
+        <>
+            {removeBgStatus === 'succeeded' && <CustomAlert title="Success" description="Image uploaded successfully" variant="success" setOpenProp />}
+            {removeBgStatus === 'failed' && <CustomAlert title="Failed" description="Something went wrong" variant="destructive" setOpenProp />}
+            {isDragOver && <CustomAlert title="Failed" description="Please select an image" variant="destructive" setOpenProp />}
+
+            {removeBgStatus === 'failed' ? (
+                <ServerError />
+            ) : (
+                <div className='w-full mx-auto flex justify-center items-center'>
+                    <Card className='w-[95%] md:w-[80%] lg:w-[60%] mt-10'>
+                        <CardHeader>
+                            <CardTitle className="text-lg md:text-2xl text-center">Remove Background</CardTitle>
+                        </CardHeader>
+                        <CardContent>
+                            {removeBgStatus === 'idle' ? (
+                                <div className="flex flex-col space-y-4 my-2 items-center ">
+                                    {!hide && (
+                                        <div className="flex flex-col items-center space-y-2 text-sm md:text-base">
+                                            <p>Before uploading the image choose the image type if both are not working use other</p>
+                                            <div className='grid grid-cols-2 md:grid-cols-3 gap-2' >
+                                                <Button variant="outline" onClick={animeHandler}>Anime</Button>
+                                                <Button variant="outline" onClick={generalHandler}>General</Button>
+                                                <Button variant="outline" onClick={otherHandler}>Other</Button>
+
+                                            </div>
+                                        </div>
+                                    )}
+                                    <Label className="text-base md:text-lg" htmlFor="image">Upload Image</Label>
+                                    <Input
+                                        name="image"
+                                        type="file"
+                                        accept="image/*"
+                                        className='w-full dark:file:text-white cursor-pointer'
+                                        onChange={(e) => { uploadHndler(e) }}
+                                    />
+                                    <div
+                                        onDrop={handleDrop}
+                                        onDragOver={handleDragOver}
+                                        className="w-full h-44 md:h-96 border-2 flex justify-center items-center rounded-md text-sm md:text-lg"
+                                    >
+                                        Drag and drop the image here
                                     </div>
                                 </div>
-                            )}
-                            <Label className="text-base md:text-lg" htmlFor="image">Upload Image</Label>
-                            <Input
-                                name="image"
-                                type="file"
-                                accept="image/*"
-                                className='w-full dark:file:text-white cursor-pointer'
-                                onChange={(e) => { uploadHndler(e) }}
-                            />
-                            <div
-                                onDrop={handleDrop}
-                                onDragOver={handleDragOver}
-                                className="w-full h-44 md:h-96 border-2 flex justify-center items-center rounded-md text-sm md:text-lg"
-                            >
-                                Drag and drop the image here
-                            </div>
-                        </div>
-                    ) : removeBgStatus === 'loading' ? (
-                        <Loader2 className="w-14 h-14 animate-spin mx-auto" />
-                    ) : removeBgStatus === 'succeeded' ? (
-                        <p className='text-center text-lg' >Image Uploaded</p>
-                    ) : removeBgStatus === 'failed' ? (
-                        <p className='text-center text-lg'>Something went wrong</p>
-                    ) : null}
-                </CardContent>
-                <CardFooter>
-                    {getRemoveBgStatus === 'succeeded' && (
-                        <div className='flex flex-col w-full space-y-4'>
-                            <p className='text-center'>Compare</p>
-                            <div className='w-full h-auto'>
-                                <ReactCompareImage leftImage={original} leftImageLabel='Original' rightImage={removeBgImage} rightImageLabel='Bg Removed' sliderLineColor='#6d28d9' />
+                            ) : removeBgStatus === 'loading' ? (
+                                <Loader2 className="w-14 h-14 animate-spin mx-auto" />
+                            ) : removeBgStatus === 'succeeded' ? (
+                                <p className='text-center text-lg' >Image Uploaded</p>
+                            ) : removeBgStatus === 'failed' ? (
+                                <p className='text-center text-lg'>Something went wrong</p>
+                            ) : null}
+                        </CardContent>
+                        <CardFooter>
+                            {getRemoveBgStatus === 'succeeded' && (
+                                <div className='flex flex-col w-full space-y-4'>
+                                    <p className='text-center'>Compare</p>
+                                    <div className='w-full h-auto'>
+                                        <ReactCompareImage leftImage={original} leftImageLabel='Original' rightImage={removeBgImage} rightImageLabel='Bg Removed' sliderLineColor='#6d28d9' />
 
-                            </div>
-                            <Button className="w-full"><a href={removeBgImage} download="removeBg.png">Download</a></Button>
-                        </div>
-                    )}
-                </CardFooter>
-            </Card>
-        </div>
+                                    </div>
+                                    <Button className="w-full"><a href={removeBgImage} download="removeBg.png">Download</a></Button>
+                                    <Button className="w-full" onClick={resetHandler}>Another Image</Button>
+                                </div>
+                            )}
+                        </CardFooter>
+                    </Card>
+                </div>
+            )}
+        </>
     )
 }
